+++ src/api/queries.rs
@@ -589,3 +589,37 @@ pub const INITIATIVE_QUERY: &str = r#"
         }
     }
 "#;
+
+pub const TEMPLATE_SEARCH_QUERY: &str = r#"
+    query TemplateSearch($filter: TemplateFilter, $first: Int) {
+        templateSearch(filter: $filter, first: $first) {
+            id
+            name
+            description
+            type
+            content
+            templateData
+            team {
+                key
+                name
+            }
+        }
+    }
+"#;
+
+pub const TEMPLATE_QUERY: &str = r#"
+    query Template($id: String!) {
+        template(id: $id) {
+            id
+            name
+            description
+            type
+            content
+            templateData
+            team {
+                key
+                name
+            }
+        }
+    }
+"#;